+++ ClientService/ClientService.csproj
@@ -43,10 +43,15 @@
   <ItemGroup>
     <Compile Include="CacheIO.cs" />
     <Compile Include="ContextProvider.cs" />
+    <Compile Include="Contracts\IProvider.cs" />
     <Compile Include="DataProvider.cs" />
+    <Compile Include="Data\CustomSizeProvider.cs" />
+    <Compile Include="Data\SizeGroupProvider.cs" />
+    <Compile Include="Data\SizeProvider.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="SizeGroupRepository.cs" />
     <Compile Include="SpecsManiuplator.cs" />
+    <Compile Include="Data\SpecsProvider.cs" />
     <Compile Include="SpecsRepository.cs" />
     <Compile Include="SpecsVMProvider.cs" />
   </ItemGroup>
