+++ ClientService/Contracts/IProvider.cs
@@ -1,0 +1,11 @@
+﻿using System.Collections.Generic;
+
+namespace ClientService.Contracts
+{
+    public interface IProvider<T>
+    {
+        List<T> GetList();
+
+        List<string> GetIDs();
+    }
+}
